Revert "reverted back to png based conversion, src, config cleanup"
This reverts commit 8f998f7ccc12dfafcadbf73bda58876da65c4a2b.
