+++ main.py
@@ -16,21 +16,17 @@ def main(cfg: DictConfig) -> None:
     """
     Main entry point for running SemiF-Preprocesing pipeline.
     """
-    try:
-        cfg = OmegaConf.create(cfg)
-        log.info(f"Starting task {','.join(cfg.tasks)}")
+    cfg = OmegaConf.create(cfg)
+    log.info(f"Starting task {','.join(cfg.tasks)}")
+    
+    for tsk in cfg.tasks:
+        try:
+            task = get_method(f"{tsk}.main")
+            task(cfg)
 
-        for tsk in cfg.tasks:
-            try:
-                task = get_method(f"{tsk}.main")
-                task(cfg)
-
-            except Exception as e:
-                log.exception("Failed")
-                sys.exit(1)     # used to communicate error with bash shell
-    except Exception as e:
-        log.exception(f"Fatal error, failed: {e}")
-        sys.exit(1)     # used to communicate error with bash shell
+        except Exception as e:
+            log.exception("Failed")
+            return
 
 
 if __name__ == "__main__":
